+++ cache.h
@@ -25,7 +25,7 @@ int
 waveform_db_delete (char const *fname);
 
 int
-waveform_db_read (char const *fname, float *buffer, int buffer_len, int *channels);
+waveform_db_read (char const *fname, short *buffer, int buffer_len, int *channels);
 
 void
-waveform_db_write (char const *fname, float *buffer, int buffer_len, int channels, int compression);
+waveform_db_write (char const *fname, short *buffer, int buffer_len, int channels, int compression);
